services/curation_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai>=1.0.0
openai-swarm
pdf2image>=1.17.0
Pillow>=10.2.0
python-dotenv>=1.0.1